+++ Runtime/CompositionSplash.cs
@@ -1,8 +1,17 @@
+using System;
 using System.Collections;
 using Unity.XR.CompositionLayers.Extensions;
+using Unity.XR.CompositionLayers.Services;
 using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.SceneManagement;
+
+#if UNITY_6000_0_OR_NEWER
+using AsyncReturnType = UnityEngine.Awaitable;
+#else
+using System.Threading.Tasks;
+using AsyncReturnType = System.Threading.Tasks.Task;
+#endif
 
 namespace Unity.XR.CompositionLayers
 {
@@ -15,16 +24,19 @@ namespace Unity.XR.CompositionLayers
         /// The CompositionLayer component that will be used to display the splash screen.
         /// </summary>
         [Header("References")]
+        [Tooltip("The CompositionLayer component that will be used to display the splash screen.")]
         public CompositionLayer compositionLayer;
 
         /// <summary>
         /// The TexturesExtension component attached to the CompositionLayer to control the splash screen texture.
         /// </summary>
+        [Tooltip("The TexturesExtension component attached to the CompositionLayer to control the splash screen texture.")]
         public TexturesExtension splashTexture;
 
         /// <summary>
         /// The ColorScaleBiasExtension component attached to the CompositionLayer to control the splash screen alpha.
         /// </summary>
+        [Tooltip("The ColorScaleBiasExtension component attached to the CompositionLayer to control the splash screen alpha.")]
         public ColorScaleBiasExtension colorScaleBias;
 
         bool m_MainCameraRendered = false;
@@ -33,19 +45,41 @@ namespace Unity.XR.CompositionLayers
 
         Camera m_MainCamera;
 
+        AsyncReturnType splashScreenAwaitable;
+
+        bool passthroughInitialized = false;
+
         void Awake()
         {
             if (!Validate()) return;
 
             m_pref = CompositionLayersRuntimeSettings.Instance;
+            if (splashScreenAwaitable != null && !splashScreenAwaitable.IsCompleted)
+            {
+#if UNITY_6000_0_OR_NEWER
+                splashScreenAwaitable.Cancel();
+#else
+                splashScreenAwaitable.Dispose();
+#endif
+                splashScreenAwaitable = null;
+            }
 
-            SetupCamera();
+            SetupCamera(false);
             SetupLayer();
-            StartCoroutine(DisplaySplashScreens());
+
+            splashScreenAwaitable = DisplaySplashscreens();
         }
 
         void Update()
         {
+            // We can't spawn the layer in awake, as it is ran before OpenXRLayerProvider.Started, where the layer is registered.
+            if (CompositionLayerManager.PassthroughLayerType != null && !passthroughInitialized)
+            {
+                bool passthroughSetup = TrySetupPassthrough();
+                SetupCamera(passthroughSetup);
+                passthroughInitialized = true;
+            }
+
             Camera camera = GetMainCamera();
 
             Vector3 lerpPosition = Vector3.Lerp(compositionLayer.transform.position, GetTargetPosition(m_pref.LockToHorizon), Time.deltaTime * m_pref.FollowSpeed);
@@ -91,6 +125,48 @@ namespace Unity.XR.CompositionLayers
             }
         }
 
+        bool TrySetupPassthrough()
+        {
+            if (m_pref == null)
+                return false;
+
+            if (m_pref.BackgroundType != CompositionLayersRuntimeSettings.SplashBackgroundType.Passthrough)
+                return false;
+
+            Type passthroughType = CompositionLayerManager.PassthroughLayerType;
+            if (passthroughType == null)
+            {
+                Debug.LogWarning("Passthrough background style is selected, but no passthrough layer is available.");
+                return false;
+            }
+
+            Layers.LayerDataDescriptor passthroughDescriptor = CompositionLayerUtils.GetLayerDescriptor(passthroughType);
+            if (passthroughDescriptor == null)
+            {
+                Debug.LogWarning("Passthrough Layer Data Descriptor is not found.");
+                return false;
+            }
+
+            Layers.LayerData passthroughLayerData = CompositionLayerUtils.CreateLayerData(passthroughType);
+            if (passthroughLayerData == null)
+            {
+                Debug.LogWarning("Passthrough Layer Data is not found.");
+                return false;
+            }
+
+            CompositionLayer passthroughLayer = new GameObject("Passthrough Layer").AddComponent<CompositionLayer>();
+
+            passthroughLayer.ChangeLayerDataType(passthroughLayerData);
+            foreach (var extension in passthroughDescriptor.SuggestedExtensions)
+            {
+                if (extension.IsSubclassOf(typeof(MonoBehaviour)))
+                    passthroughLayer.gameObject.AddComponent(extension);
+            }
+
+            passthroughLayer.TryChangeLayerOrder(passthroughLayer.Order, CompositionLayerManager.GetFirstUnusedLayer(false));
+            return true;
+        }
+
         Vector3 TransformPointToNearestPointOnSphere(Vector3 point, Vector3 center, float radius)
         {
             Vector3 direction = point - center;
@@ -98,7 +174,7 @@ namespace Unity.XR.CompositionLayers
             return center + direction * radius;
         }
 
-        IEnumerator DisplaySplashScreens()
+        async AsyncReturnType DisplaySplashscreens()
         {
             // Set splash texture
             if (m_pref.SplashImage != null)
@@ -119,12 +195,12 @@ namespace Unity.XR.CompositionLayers
                 {
                     timer += Time.deltaTime;
                     SetColorScale(ColorScaleBiasLerp(timer, m_pref.FadeInDuration));
-                    yield return null;
+                    await WaitForEndOfFrame();
                 }
             }
 
             // Wait for splash duration
-            yield return new WaitForSeconds(m_pref.SplashDuration);
+            await WaitForSeconds(m_pref.SplashDuration);
 
             // Fade out
             if (m_pref.FadeOutDuration <= 0)
@@ -138,7 +214,7 @@ namespace Unity.XR.CompositionLayers
                 {
                     timer += Time.deltaTime;
                     SetColorScale(ColorScaleBiasLerp(m_pref.FadeOutDuration - timer, m_pref.FadeOutDuration));
-                    yield return null;
+                    await WaitForEndOfFrame();
                 }
             }
 
@@ -152,12 +228,15 @@ namespace Unity.XR.CompositionLayers
             colorScaleBias.Scale = scale;
         }
 
-        void SetupCamera()
+        void SetupCamera(bool passthrough)
         {
             Camera camera = GetMainCamera();
 
+            if (camera == null)
+                return;
+
             camera.clearFlags = CameraClearFlags.SolidColor;
-            camera.backgroundColor = m_pref.BackgroundColor;
+            camera.backgroundColor = passthrough ? Color.clear : m_pref.BackgroundColor;
         }
 
         Vector4 ColorScaleBiasLerp(float timer, float duration)
@@ -222,6 +301,24 @@ namespace Unity.XR.CompositionLayers
                 m_MainCamera = Camera.main;
 
             return m_MainCamera;
+        }
+
+        async AsyncReturnType WaitForEndOfFrame()
+        {
+#if UNITY_6000_0_OR_NEWER
+            await Awaitable.EndOfFrameAsync();
+#else
+            await Task.Yield();
+#endif
+        }
+
+        async AsyncReturnType WaitForSeconds(float seconds)
+        {
+#if UNITY_6000_0_OR_NEWER
+            await Awaitable.WaitForSecondsAsync(seconds);
+#else
+            await Task.Delay((int)(seconds * 1000));
+#endif
         }
     }
 }